src/data-notifier/requirements.txt
requests
schedule
urllib3
kafka-python